# api/openapi.yaml
openapi: 3.0.3
info:
  title: BeeBase Hive Service API
  description: >
    Hive management service for BeeBase, an open-source backend for a
    beekeeper management application split into microservices. Access
    tokens are issued by auth-service (see its OpenAPI spec); this
    service verifies them against auth-service's public key and never
    trusts a userID from anywhere but a verified token.

    Every hive belongs to an apiary owned by apiary-service. Creating a
    hive requires that the caller own the referenced apiary; this service
    confirms that by forwarding the caller's own access token to
    apiary-service's GET /api/v1/apiaries/{id} and trusting its answer -
    it never queries apiary ownership itself.

    Photos and other files are uploaded directly to media-service (see its
    OpenAPI spec) independently of any hive - upload doesn't require one
    to exist yet, and this service never receives file bytes. `images` on
    a hive is the set of media IDs this service itself stores and returns
    - the source of truth for what's referenced, not a media-service
    round trip - settable via POST (reference immediately on creation) or
    PUT (replace the set wholesale). Every id is verified against
    media-service (a plain ownership check, not a mutating call) before
    being accepted; an id that doesn't exist, was deleted, or belongs to
    a different user is rejected. There's no exclusivity between
...
    every file it references).
  version: 0.7.0
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/hives:
    post:
      tags: [hives]
      summary: Create a hive under one of the caller's own apiaries
      description: >
        Rejected with 404 apiary_not_found if apiary_id doesn't exist or
        doesn't belong to the caller — verified live against
        apiary-service, not trusted from the request. images, if given, is
        referenced immediately: each id is verified against media-service
        (it must belong to the caller) before the hive is created - an
        unknown id or someone else's is rejected and no hive is created.
        Omitting images (or sending an empty array) creates a hive with no
        photos; more can be added afterward via PUT.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateHiveRequest" }
      responses:
        "201":
          description: Hive created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HiveResponse" }
        "400":
          description: Request failed validation (including a missing or malformed apiary_id), or media limit reached.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/ValidationErrorResponse" }
                  - { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: >
            Either the Free plan creation limit was reached (code
            hiveLimitReached: the account already owns 5 hives - a
            per-user, account-wide quota, independent of how many apiaries
            they're spread across), or apiary_id itself is not currently
            writable under Free (code parent_resource_pro_locked). Pro has
            no such limit.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: apiary_id doesn't exist, or belongs to a different user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    get:
      tags: [hives]
      summary: List the authenticated user's hives
      description: >
        Returns only hives owned by the caller — never another user's — as
        a page of results. Omitting page/limit falls back to page=1,
        limit=20. Each item's `images` reflects that hive's actual
        attached media ids - it's a local column now, not a per-item
        media-service round trip, so listing carries no extra cost to
        include it. Each item's `writable` is resolved once for the whole
        page (at most one extra call to apiary-service, regardless of how
        many distinct apiaries the page spans) and reflects the caller's
        complete set of hives, never just the current page or search
        result.
      parameters:
        - name: page
          in: query
          description: 1-indexed page number.
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          description: Rows per page.
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: search
          in: query
          description: Optional search filter (matched case-insensitively against name and notes). Minimum 3 characters.
          required: false
          schema: { type: string, minLength: 3 }
        - name: sortOrder
          in: query
          description: >
            Optional sort direction. Omitted keeps the endpoint's default
            order (created_at ascending, id ascending); when given ("asc"
            or "desc"), results are instead ordered by creation date
            (created_at) in that direction, with id tied to the same
            direction as a stable secondary sort.
          required: false
          schema: { type: string, enum: [asc, desc] }
        - name: needsInspection
          in: query
          description: >
            When exactly "true", restricts results to hives that
            currently need inspection: never inspected, or last
            inspected more than the configured warning threshold ago
            (see GET /api/v1/statistics/needs-attention on
            statistics-service, which reports the same threshold and the
            same count). Any other value (including "false", or
            omitting it) leaves results unfiltered. Computed live
            against inspection-service on every request - the caller
            never supplies the threshold itself.
          required: false
          schema: { type: string, enum: [true, false] }
      responses:
        "200":
          description: A page of the caller's hives.
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/HiveResponse" }
                  pagination: { $ref: "#/components/schemas/Pagination" }
        "400":
          description: page, limit, search, or sortOrder is invalid.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationErrorResponse" }
        "401":
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: A required upstream service (inspection-service) failed or was unreachable while computing needs_inspection.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/hives/{hiveId}:
    parameters:
      - name: hiveId
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [hives]
      summary: Get one of the authenticated user's hives
      description: >
        Returns 404 both when hiveId doesn't exist and when it belongs to
        a different user — the two cases are indistinguishable to the
        caller, by design.
      responses:
        "200":
          description: The hive.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HiveResponse" }
        "400":
          description: hiveId is not a valid UUID.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: No such hive, or it belongs to a different user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    put:
      tags: [hives]
      summary: Replace a hive's editable fields
      description: >
        Full replacement (PUT semantics), not a partial patch, for name and
        notes. A hive can't be moved to a different apiary —
        there's no apiary_id here. `images` is the one field exempt from
        full-replacement semantics: omitting it (or sending JSON null)
        leaves the current set untouched, while sending an array
        (including an empty one) replaces it wholesale - every id is
        verified against media-service (it must belong to the caller)
        before anything changes. Dropping an id from the list only
        removes the reference; it does not delete the underlying file.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateHiveRequest" }
      responses:
        "200":
          description: The updated hive.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HiveResponse" }
        "400":
          description: >
            hiveId is not a valid UUID, the request failed validation,
            media limit reached, or images contains an id that doesn't belong to the caller.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/ValidationErrorResponse" }
                  - { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: >
            The hive is currently read-only under the caller's Free
            entitlement: either the hive itself is outside the first 5 in
            its apiary (code resource_pro_locked), or its parent apiary is
            itself not the caller's writable one (code
            parent_resource_pro_locked). Upgrading to Pro, or a promotion
            (e.g. a currently-writable sibling is deleted), unblocks this
            immediately.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: No such hive, or it belongs to a different user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    delete:
      tags: [hives]
      summary: Delete a hive
      description: Hard delete after cascading deletion of the hive's inspections, harvests, reminders, and attached media.
      responses:
        "204":
          description: The hive and its descendants are deleted.
        "400":
          description: hiveId is not a valid UUID.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: No such hive, or it belongs to a different user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/apiaries/{apiaryId}/hives:
    parameters:
      - name: apiaryId
        in: path
        required: true
        description: Apiary ID.
        schema: { type: string, format: uuid }

    get:
      tags: [hives]
      summary: List the authenticated user's hives for one apiary
      description: >
        Returns only hives owned by the caller belonging to this apiary, as a
        page of results. Rejected with 404 apiary_not_found if apiaryId doesn't
        exist or doesn't belong to the caller — verified live against
        apiary-service. Omitting page/limit falls back to page=1, limit=20.
        Each item's images reflects that hive's actual attached media ids.
      parameters:
        - name: page
          in: query
          description: 1-indexed page number.
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          description: Rows per page.
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: search
          in: query
          description: Optional search filter (matched case-insensitively against name and notes). Minimum 3 characters.
          required: false
          schema: { type: string, minLength: 3 }
        - name: sortOrder
          in: query
          description: >
            Optional sort direction. Omitted keeps the endpoint's default
            order (created_at ascending, id ascending); when given ("asc"
            or "desc"), results are instead ordered by creation date
            (created_at) in that direction, with id tied to the same
            direction as a stable secondary sort.
          required: false
          schema: { type: string, enum: [asc, desc] }
        - name: needsInspection
          in: query
          description: >
            When exactly "true", restricts results to hives in this
            apiary that currently need inspection. See the same
            parameter on GET /api/v1/hives for the full rule.
          required: false
          schema: { type: string, enum: [true, false] }
      responses:
        "200":
          description: A page of the caller's hives for this apiary.
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/HiveResponse" }
                  pagination: { $ref: "#/components/schemas/Pagination" }
        "400":
          description: apiaryID is not a valid UUID, or page, limit, search, or sortOrder is invalid.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/ValidationErrorResponse" }
                  - { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: apiaryID doesn't exist, or belongs to a different user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/hives/{hiveId}/queen:
    get:
      tags: [queens]
      summary: Get the current queen of a hive
      description: Returns the currently active queen (removed_at is null) for the specified hive.
      parameters:
        - name: hiveId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Current queen found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueenResponse" }
        "404":
          description: Hive or queen not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/hives/{hiveId}/queens:
    get:
      tags: [queens]
      summary: List queen history for a hive
      parameters:
        - name: hiveId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Queen history returned.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/QueenResponse" }
        "404":
          description: Hive not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      tags: [queens]
      summary: Register a new queen for a hive
      description: >
        Registers a queen into the hive's strict chronological queen lifecycle chain.
        Derives lifecycle boundaries automatically and transactionally.
        If inserted after the current queen, closes the previous queen and becomes the new current queen.
        If inserted retroactively between existing queens or before the oldest queen, splits the affected neighboring boundaries.
      parameters:
        - name: hiveId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateQueenRequest" }
      responses:
        "201":
          description: Queen registered.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueenResponse" }
        "400":
          description: Validation error (e.g. year invalid, missing introducedAt).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationErrorResponse" }
        "409":
          description: Conflict (e.g. duplicate_introduced_at).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An access token issued by auth-service's /api/v1/auth/login (or /register or /refresh).
  schemas:
...
    CreateHiveRequest:
      type: object
      required: [apiaryId, name]
      properties:
        apiaryId: { type: string, format: uuid, description: Must be an apiary owned by the caller. }
        name: { type: string, maxLength: 200 }
        notes: { type: string, maxLength: 2000 }
        images:
          type: array
          maxItems: 5
          items: { type: string, format: uuid }
          description: >
            Media IDs (returned by media-service's upload endpoint) to
            reference immediately. Each must belong to the caller; omit
            this field (or send an empty array) to create a hive with no
            photos. Maximum 5 photos allowed.
    UpdateHiveRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 200 }
        notes: { type: string, maxLength: 2000 }
        images:
          type: array
          maxItems: 5
          items: { type: string, format: uuid }
          description: >
            The desired final set of media IDs (returned by media-service's
            upload endpoint) referenced by this hive. Omit this field (or
            send null) to leave the current set untouched; send an array
            (including an empty one) to replace it wholesale. Each id must
            belong to the caller. Maximum 5 photos allowed. Dropping an id
            only removes the reference - it does not delete the underlying file.
    HiveImage:
      type: object
      properties:
        id: { type: string, format: uuid }
        imageUrl:
          type: string
          format: uri
          description: >
            The authenticated URL to fetch or display this image's
            content from (media-service's GET /media/{id}/download,
            reverse-proxied by the gateway). Built fresh on every
            response - if the underlying file ever changes, this URL's
            id stays the same but its content does not, so clients
            should not cache a fetched image keyed only by this URL
            without also checking for staleness.
    HiveResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        apiaryId: { type: string, format: uuid }
        name: { type: string }
        notes: { type: string }
        images:
          type: array
          items: { $ref: "#/components/schemas/HiveImage" }
          description: >
            The media currently attached to this hive, in every response
            including list items (see GET /hives) - never null, empty
            when there are no photos.
        writable:
          type: boolean
          description: >
            Whether the caller can currently edit this hive, create
            inspections/harvests under it, or otherwise mutate it or its
            descendants. Always true on a Pro subscription. Under Free,
            true only when this hive's parent apiary is itself writable
            (see ApiaryResponse.writable on apiary-service) AND this hive
            ranks among the caller's first 5 hives account-wide by
            creation order (FreeMaxHives is a per-user quota, not scoped
            to any single apiary - a hive sitting in a different,
            currently read-only apiary still counts against this
            ranking, even though it can never itself be writable) - every
            other hive remains fully readable (including its
            inspections/harvests/media history) but rejects writes with
            403 resource_pro_locked (this hive itself is outside the
            limit) or parent_resource_pro_locked (its apiary is) until
            it's promoted (e.g. a currently-writable sibling is deleted)
            or the caller upgrades to Pro. Deleting a hive is never
            affected by this flag.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        currentQueen:
          $ref: "#/components/schemas/QueenResponse"
          description: >
            The currently active queen for this hive, if any (removed_at is null).
            Omitted when the hive currently has no active queen.
    Pagination:
      type: object
      description: Metadata describing the requested page relative to the full result set.
      required: [page, limit, total, totalPages, hasNext, hasPrevious]
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, description: Total matching rows across all pages., example: 125 }
        totalPages: { type: integer, example: 7 }
        hasNext: { type: boolean }
        hasPrevious: { type: boolean }
    QueenResponse:
      type: object
      required: [id, hiveId, year, markingColor, markingColorHex, introducedAt, notes, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        hiveId: { type: string, format: uuid }
        year: { type: integer, example: 2026 }
        markingColor: { type: string, enum: [blue, white, yellow, red, green], example: white }
        markingColorHex: { type: string, example: "#F8F9FA" }
        markedAt: { type: string, format: date-time, nullable: true }
        introducedAt: { type: string, format: date-time, example: "2026-04-10T00:00:00Z" }
        removedAt: { type: string, format: date-time, nullable: true }
        notes: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateQueenRequest:
      type: object
      required: [year, introducedAt]
      properties:
        year: { type: integer, example: 2026 }
        introducedAt: { type: string, format: date-time, example: "2026-04-10T00:00:00Z" }
        markedAt: { type: string, format: date-time, nullable: true }
        notes: { type: string }
...
    ErrorResponse:
      type: object
      description: >
        "code" is the stable, machine-readable key: build any localized
        message off of it, not off "message" (an English fallback for logs
        and manual debugging, not meant to be shown to end users).
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - hive_not_found
                - apiary_not_found
                - invalid_hive_id
                - missing_authorization
                - invalid_access_token
                - invalid_body
                - internal_error
                - media_limit_reached
                - hive_limit_reached
                - hive_name_exists
                - resource_pro_locked
                - parent_resource_pro_locked
                - queen_not_found
                - invalid_queen_id
                - active_queen_exists
                - no_active_queen
                - queen_not_latest
                - duplicate_introduced_at
              example: hive_not_found
            message: { type: string, example: hive not found }
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, enum: [validation_error] }
            message: { type: string, example: request validation failed }
            fields:
              type: object
              description: >
                Field name to error code. Each value is a stable key (not a
                human-readable message) meant to be localized client-side.
                "images" carries "images_invalid" when an entry isn't a
                well-formed UUID, or "image_not_found" when an entry is a
                well-formed UUID that doesn't belong to the caller
                (whether because it doesn't exist, was deleted, or
                belongs to a different user - these are deliberately
                indistinguishable, the same non-leaking convention
                hive_not_found already follows).
              additionalProperties:
                type: string
                enum:
                  - apiary_id_required
                  - apiary_id_invalid
                  - name_required
                  - name_too_long
                  - notes_too_long
                  - images_invalid
                  - image_not_found
                  - invalid_page
                  - invalid_limit
                  - invalid_search
                  - invalid_sort_order
                  - year_required
                  - year_invalid
                  - introduced_at_required
                  - timeline_invalid
              example: { apiary_id: "apiary_id_invalid" }